pyproject.toml
[tool.poetry]
name = "final-assignment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pygame = "^2.0.1"
pygame-widgets = "^0.4.0"
Pillow = "^10.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
